// Weekly/May/3W-7D/ecommerce/src/app/interfaces/product.interface.ts
export interface Product {
  id:                 number;
  title:              string;
  description:        string;
  price:              number;
  discountPercentage: number;
  rating:             number;
  stock:              number;
  brand:              string;
  category:           string;
  thumbnail:          string;
  images:             string[];
}

export interface ProductResponse {
  products: Product[];
  total:    number;
  skip:     number;
  limit:    number;
}

export interface CartItem extends Product {
  amount:   number
  totalPrice: number
}